Support Python 3.13

+++ pytams/xmlutils.py
@@ -14,7 +14,7 @@ _logger = logging.getLogger(__name__)
 def oneliner_ndarray():
     """Force ndarray print on a single line temporarily."""
     oldoptions = np.get_printoptions()
-    np.set_printoptions(linewidth=np.inf, precision=16)
+    np.set_printoptions(linewidth=np.inf, precision=12)
     yield
     np.set_printoptions(**oldoptions)
